pyproject.toml
[tool.poetry]
name = "langserveapp"
version = "0.1.0"
description = ""
authors = ["Beta Lin <sb2828sb@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = ">2"
langchain = "^0.2.7"
langchain-openai = "^0.1.16"
qdrant-client = "^1.10.1"
langchain-community = "^0.2.7"
pypdf = "^4.2.0"
langchain-cli = "^0.0.25"
line-bot-sdk = "^3.11.0"
httpx = "^0.27.0"
python-dotenv = "^0.19.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
